# swagger/index.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Calendar booking API v6",
    "description": "Calendar booking API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "summary": "Create new user",
        "description": "Create new user",
        "parameters": [
          {
            "name": "user",
            "in": "body",
            "description": "User with new values of properties",
            "schema": {
              "$ref": "#/definitions/CreateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
...
    "/residences": {
      "get": {
        "tags": [
          "Residences"
        ],
        "summary": "Get all residences in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Residence"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Residence"
        ],
        "summary": "Create new residence",
        "description": "Create new residence",
        "parameters": [
          {
            "name": "residence",
            "in": "body",
            "description": "Residence with new values of properties",
            "schema": {
              "$ref": "#/definitions/Residence"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Residence"
            }
          }
        }
      }
    },
...
    "/apartments": {
      "get": {
        "tags": [
          "Apartments"
        ],
        "summary": "Get all apartments in system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Apartment"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Apartment"
        ],
        "summary": "Create new apartment",
        "description": "Create new apartment",
        "parameters": [
          {
            "name": "apartment",
            "in": "body",
            "description": "Apartment with new values of properties",
            "schema": {
              "$ref": "#/definitions/Apartment"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/Apartment"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "required": [
        "username",
        "email",
        "firstname",
        "lastname",
        "apartment",
        "residence"
      ],
      "properties": {
        "_id": {
          "type": "string",
          "uniqueItems": true
        },
        "hash": {
          "type": "string",
          "required": true
        },
        "email": {
          "type": "string",
          "required": true
        },
        "username": {
          "type": "string",
          "required": true
        },
        "firstname": {
          "type": "string",
          "required": true
        },
        "lastname": {
          "type": "string",
          "required": true
        },
        "apartment": {
          "type": "string",
          "uniqueItems": true
        },
        "residence": {
          "type": "string",
          "uniqueItems": true
        }
      }
    },
...
    "CreateUser": {
      "required": [
        "username",
        "email",
        "password",
        "firstname",
        "lastname",
        "apartment",
        "residence"
      ],
      "properties": {
        "email": {
          "type": "string",
          "required": true
        },
        "username": {
          "type": "string",
          "required": true
        },
        "firstname": {
          "type": "string",
          "required": true
        },
        "lastname": {
          "type": "string",
          "required": true
        },
        "apartment": {
          "type": "string",
          "required": true,
          "uniqueItems": true
        },
        "residence": {
          "type": "string",
          "required": true,
          "uniqueItems": true
        }
      }
    },
    "Residence": {
      "required": [
        "name",
        "address"
      ],
      "properties": {
        "name": {
          "type": "string",
          "required": true
        },
        "address": {
          "type": "string",
          "required": true
        }
      }
    },
...
    "Apartment": {
      "required": [
        "name",
        "residence"
      ],
      "properties": {
        "name": {
          "type": "string",
          "required": true
        },
        "residence": {
          "type": "string",
          "required": true
        }
      }
    },
...
  }
}